refactor: remove `miden-abi-conversion` crate and move its code to
`frontend-wasm` crate.
