+++ frontend-wasm/src/code_translator/mod.rs
@@ -25,7 +25,7 @@ use wasmparser::{MemArg, Operator};
 
 use crate::{
     error::{WasmError, WasmResult},
-    miden_abi::adapt_call,
+    miden_abi::{is_miden_sdk_module, transform::transform_miden_abi_call},
     module::{
         func_translation_state::{ControlStackFrame, ElseData, FuncTranslationState},
         function_builder_ext::FunctionBuilderExt,
@@ -649,24 +649,33 @@ fn translate_call(
     let wasm_sig = module_state.signature(function_index);
     let num_wasm_args = wasm_sig.params().len();
     let args = func_state.peekn(num_wasm_args);
-    let results = adapt_call(func_id, &args, builder, span, diagnostics);
-    assert_eq!(
-        results.len(),
-        wasm_sig.results().len(),
-        "Adapted function call results quantity are not the same as the original Wasm function \
-         results quantity"
-    );
-    assert_eq!(
-        wasm_sig.results().iter().map(|p| &p.ty).collect::<Vec<&Type>>(),
-        results
-            .iter()
-            .map(|r| builder.data_flow_graph().value_type(*r))
-            .collect::<Vec<&Type>>(),
-        "Adapted function call result types are not the same as the original Wasm function result \
-         types"
-    );
-    func_state.popn(num_wasm_args);
-    func_state.pushn(&results);
+    if is_miden_sdk_module(func_id.module.as_symbol()) {
+        // Miden SDK function call, transform the call to the Miden ABI if needed
+        let results = transform_miden_abi_call(func_id, args, builder, span, diagnostics);
+        assert_eq!(
+            wasm_sig.results().len(),
+            results.len(),
+            "Adapted function call results quantity are not the same as the original Wasm \
+             function results quantity"
+        );
+        assert_eq!(
+            wasm_sig.results().iter().map(|p| &p.ty).collect::<Vec<&Type>>(),
+            results
+                .iter()
+                .map(|r| builder.data_flow_graph().value_type(*r))
+                .collect::<Vec<&Type>>(),
+            "Adapted function call result types are not the same as the original Wasm function \
+             result types"
+        );
+        func_state.popn(num_wasm_args);
+        func_state.pushn(&results);
+    } else {
+        // no transformation needed
+        let call = builder.ins().call(func_id, &args, span);
+        let results = builder.inst_results(call);
+        func_state.popn(num_wasm_args);
+        func_state.pushn(results);
+    };
     Ok(())
 }
