--- miden-abi-conversion/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "miden-abi-conversion"
-description = "Miden ABI conversion library"
-version.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-repository.workspace = true
-categories.workspace = true
-keywords.workspace = true
-license.workspace = true
-readme.workspace = true
-edition.workspace = true
-
-[dependencies]
-miden-hir-type.workspace = true
-miden-diagnostics.workspace = true
-rustc-hash.workspace = true
-
-[dev-dependencies]
