--- miden-abi-conversion/src/adapter/mod.rs
@@ -1,13 +1,0 @@
-//! Adapter generation for converting between ad-hoc Miden ABI and C ABI
-
-// use miden_diagnostics::{DiagnosticsHandler, SourceSpan};
-
-// pub enum AdapterError {}
-
-// pub fn generate_adapter(
-//     // builder: &mut FunctionBuilderExt,
-//     span: SourceSpan,
-//     diagnostics: &DiagnosticsHandler,
-// ) -> Result<(), AdapterError> {
-//     Ok(())
-// }
